Implement a Rust main
This patch adds an executable written in Rust that can link to the existing C code.
It also adds logic to build the entire project via Cargo and link it to said main.
It is probably best to move forward with Cargo as the build system, and slowly
rewrite components in Rust.