Borrow parsing infrastructure from TIR and write grammar
This commit adds a basic Lalrpop grammar for the language, there is a single conflict in the gramar that must
be resovled, and AST/Term construction must still be written. Currently we expose two entry points into the
parser one for full programs, and one for commands (which can be used to parse, type check and run a single
command).